Optimized trading algo for tradingview to send webhook requests to the flask api
