+++ .env
@@ -7,14 +7,14 @@ ALPACA_BASE_URL=https://paper-api.alpaca.markets  # use live URL when you’re r
 WEBHOOK_SECRET=super_strong_secret_here
 
 # === Trading Settings ===
-ALLOWED_SYMBOLS=SPY,QQQ
-DAILY_MAX_LOSS=1000          # bot shuts down after losing this much in a day
-MAX_TRADES_PER_DAY=20        # cap number of trades to prevent overtrading
-MAX_CONCURRENT_POSITIONS=2   # only allow this many open positions at once
+#ALLOWED_SYMBOLS=SPY,QQQ
+#DAILY_MAX_LOSS=1000
+#MAX_TRADES_PER_DAY=20
+#MAX_CONCURRENT_POSITIONS=2
 
 # === Risk Settings ===
-STOP_PCT=0.01                # default 1% stop if no ATR sizing applied
-RISK_PER_TRADE=0.02          # risk 2% of portfolio per trade (ATR sizing later)
+#STOP_PCT=0.01                # default 1% stop if no ATR sizing applied
+#RISK_PER_TRADE=0.02          # risk 2% of portfolio per trade (ATR sizing later)
 
 # === Logging ===
-LOG_LEVEL=INFO
+#LOG_LEVEL=INFO
