+++ .gitignore
@@ -1,1 +1,1 @@
-.env
+*.env
